+++ packages/widget/src/stores/form/FormUpdater.tsx
@@ -1,21 +1,30 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 import { useAccount } from '../../hooks/useAccount.js';
 import { useChains } from '../../hooks/useChains.js';
 import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.js';
-import type { DefaultValues, FormFieldNames } from './types.js';
+import { useBookmarkActions } from '../../stores/bookmarks/useBookmarkActions.js';
+import { formDefaultValues } from '../../stores/form/createFormStore.js';
+import { useSendToWalletActions } from '../../stores/settings/useSendToWalletStore.js';
+import type { DefaultValues } from './types.js';
 import { useFieldActions } from './useFieldActions.js';
 
 export const FormUpdater: React.FC<{
-  defaultValues: Partial<DefaultValues>;
-}> = ({ defaultValues }) => {
-  const { fromChain, toChain } = useWidgetConfig();
+  reactiveFormValues: Partial<DefaultValues>;
+}> = ({ reactiveFormValues }) => {
+  const { fromChain, toChain, toAddress } = useWidgetConfig();
   const { account } = useAccount();
   const { chains } = useChains();
-  const { isTouched, resetField, setFieldValue, getFieldValues } =
-    useFieldActions();
-  const previousDefaultValues = useRef(defaultValues);
+  const { setSendToWallet } = useSendToWalletActions();
+  const { setSelectedBookmark } = useBookmarkActions();
 
+  const {
+    isTouched,
+    resetField,
+    setFieldValue,
+    getFieldValues,
+    setUserAndDefaultValues,
+  } = useFieldActions();
   // Set wallet chain as default if no chains are provided by config and if they were not changed during widget usage
   useEffect(() => {
     const chainAllowed =
@@ -50,21 +59,42 @@ export const FormUpdater: React.FC<{
   // Makes widget config options reactive to changes
   // should update userValues when defaultValues updates and includes additional logic for chains
   useEffect(() => {
-    (Object.keys(defaultValues) as FormFieldNames[]).forEach((key) => {
-      if (previousDefaultValues.current[key] !== defaultValues[key]) {
-        const value =
-          defaultValues[key] ||
-          // set the chain to the current user one if it is not present in the config
-          (key === 'fromChain' || key === 'toChain'
-            ? account.chainId || ''
-            : '');
-        setFieldValue(key, value);
-        resetField(key, { defaultValue: value });
-      }
-    });
-    previousDefaultValues.current = defaultValues;
+    if (reactiveFormValues.toAddress) {
+      setSendToWallet(true);
+    }
+
+    setSelectedBookmark(toAddress);
+
+    setUserAndDefaultValues(
+      accountForChainId(reactiveFormValues, account.chainId),
+    );
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [defaultValues, getFieldValues, resetField, setFieldValue]);
+  }, [
+    toAddress,
+    reactiveFormValues,
+    getFieldValues,
+    resetField,
+    setFieldValue,
+    setUserAndDefaultValues,
+    setSendToWallet,
+    setSelectedBookmark,
+  ]);
 
   return null;
+};
+
+const accountForChainId = (
+  defaultValues: Partial<DefaultValues>,
+  chainId?: number,
+) => {
+  const result: Partial<DefaultValues> = { ...defaultValues };
+  for (const key in result) {
+    const k = key as keyof DefaultValues;
+    if (result[k] === formDefaultValues[k]) {
+      if ((k === 'fromChain' || k === 'toChain') && chainId) {
+        result[k] = chainId;
+      }
+    }
+  }
+  return result;
 };